[SOIN] Reorganise les ressources OIDC
