+++ back/src/api/configurationServeur.ts
@@ -1,6 +1,6 @@
 import { FournisseurChemin } from './fournisseurChemin';
 import { Middleware } from './middleware';
-import { AdaptateurOIDC } from './adaptateurOIDC';
+import { AdaptateurOIDC } from './oidc/adaptateurOIDC';
 
 export type ConfigurationServeur = {
   fournisseurChemin: FournisseurChemin;
